requirements-dev.txt
ansible-core==2.14.5 ; python_version >= "3.10" and python_version < "4.0"
ansible==7.5.0 ; python_version >= "3.10" and python_version < "4.0"
ansicon==1.89.0 ; python_version >= "3.10" and python_version < "4" and platform_system == "Windows"
arrow==1.2.3 ; python_version >= "3.10" and python_version < "4"
asgiref==3.6.0 ; python_version >= "3.10" and python_version < "4.0"
astroid==2.15.4 ; python_version >= "3.10" and python_version < "4.0"
black==23.3.0 ; python_version >= "3.10" and python_version < "4.0"
blessed==1.20.0 ; python_version >= "3.10" and python_version < "4"
certifi==2023.5.7 ; python_version >= "3.10" and python_version < "4.0"
cffi==1.15.1 ; python_version >= "3.10" and python_version < "4.0"
charset-normalizer==3.1.0 ; python_version >= "3.10" and python_version < "4.0"
click==8.1.3 ; python_version >= "3.10" and python_version < "4.0"
colorama==0.4.6 ; python_version >= "3.10" and python_version < "4.0" and platform_system == "Windows" or python_version >= "3.10" and python_version < "4.0" and sys_platform == "win32"
crispy-bulma==0.9.0 ; python_version >= "3.10" and python_version < "4.0"
cryptography==40.0.2 ; python_version >= "3.10" and python_version < "4.0"
debugpy==1.6.7 ; python_version >= "3.10" and python_version < "4.0"
dill==0.3.6 ; python_version >= "3.10" and python_version < "4.0"
django-cleanup==7.0.0 ; python_version >= "3.10" and python_version < "4.0"
django-crispy-forms==2.0 ; python_version >= "3.10" and python_version < "4.0"
django-dbbackup==4.0.2 ; python_version >= "3.10" and python_version < "4.0"
django-extensions==3.2.1 ; python_version >= "3.10" and python_version < "4.0"
django-filter==23.2 ; python_version >= "3.10" and python_version < "4.0"
django-picklefield==3.1 ; python_version >= "3.10" and python_version < "4"
django-pictures==1.0.2 ; python_version >= "3.10" and python_version < "4.0"
django-prometheus==2.3.1 ; python_version >= "3.10" and python_version < "4.0"
django-q==1.3.9 ; python_version >= "3.10" and python_version < "4"
django==4.2.1 ; python_version >= "3.10" and python_version < "4.0"
duckdb==0.7.1 ; python_version >= "3.10" and python_version < "4.0"
gunicorn==20.1.0 ; python_version >= "3.10" and python_version < "4.0"
idna==3.4 ; python_version >= "3.10" and python_version < "4.0"
isort==5.12.0 ; python_version >= "3.10" and python_version < "4.0"
jinja2==3.1.2 ; python_version >= "3.10" and python_version < "4.0"
jinxed==1.2.0 ; python_version >= "3.10" and python_version < "4" and platform_system == "Windows"
lazy-object-proxy==1.9.0 ; python_version >= "3.10" and python_version < "4.0"
markupsafe==2.1.2 ; python_version >= "3.10" and python_version < "4.0"
mccabe==0.7.0 ; python_version >= "3.10" and python_version < "4.0"
mypy-extensions==1.0.0 ; python_version >= "3.10" and python_version < "4.0"
mysqlclient==2.1.1 ; python_version >= "3.10" and python_version < "4.0"
numpy==1.24.3 ; python_version < "4.0" and python_version >= "3.10"
packaging==23.1 ; python_version >= "3.10" and python_version < "4.0"
pandas==1.5.3 ; python_version >= "3.10" and python_version < "4.0"
pathspec==0.11.1 ; python_version >= "3.10" and python_version < "4.0"
pillow==9.5.0 ; python_version >= "3.10" and python_version < "4.0"
platformdirs==3.5.0 ; python_version >= "3.10" and python_version < "4.0"
prometheus-client==0.16.0 ; python_version >= "3.10" and python_version < "4.0"
pyarrow==12.0.0 ; python_version >= "3.10" and python_version < "4.0"
pycparser==2.21 ; python_version >= "3.10" and python_version < "4.0"
pylance==0.3.19 ; python_version >= "3.10" and python_version < "4.0"
pylint-django==2.5.3 ; python_version >= "3.10" and python_version < "4.0"
pylint-plugin-utils==0.7 ; python_version >= "3.10" and python_version < "4.0"
pylint==2.17.4 ; python_version >= "3.10" and python_version < "4.0"
python-dateutil==2.8.2 ; python_version >= "3.10" and python_version < "4.0"
python-dotenv==1.0.0 ; python_version >= "3.10" and python_version < "4.0"
pytube @ git+https://github.com/julySteuer/pytube@master ; python_version >= "3.10" and python_version < "4.0"
pytz==2023.3 ; python_version >= "3.10" and python_version < "4.0"
pyyaml==6.0 ; python_version >= "3.10" and python_version < "4.0"
redis==3.5.3 ; python_version >= "3.10" and python_version < "4"
requests==2.30.0 ; python_version >= "3.10" and python_version < "4.0"
resolvelib==0.8.1 ; python_version >= "3.10" and python_version < "4.0"
setuptools==67.7.2 ; python_version >= "3.10" and python_version < "4.0"
six==1.16.0 ; python_version >= "3.10" and python_version < "4"
sqlparse==0.4.4 ; python_version >= "3.10" and python_version < "4.0"
tomli==2.0.1 ; python_version >= "3.10" and python_version < "3.11"
tomlkit==0.11.8 ; python_version >= "3.10" and python_version < "4.0"
tqdm==4.65.0 ; python_version >= "3.10" and python_version < "4.0"
typing-extensions==4.5.0 ; python_version >= "3.10" and python_version < "3.11"
tzdata==2023.3 ; python_version >= "3.10" and python_version < "4.0" and sys_platform == "win32"
urllib3==2.0.2 ; python_version >= "3.10" and python_version < "4.0"
wcwidth==0.2.6 ; python_version >= "3.10" and python_version < "4"
wrapt==1.15.0 ; python_version >= "3.10" and python_version < "4.0"
